restful + json格式请求体

+++ models/admin_log.go
@@ -100,7 +100,7 @@ func (m *AdminLog) FindById(id int) (adminLog AdminLog, err error) {
 	return
 }
 
-func (m *AdminLog) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []AdminLog, total int, err error) {
+func (m *AdminLog) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []AdminLog, total int64, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
