+++ models/category.go
@@ -109,7 +109,7 @@ func (m *Category) FindById(id int) (category Category, err error) {
 	return
 }
 
-func (m *Category) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []Category, total int, err error) {
+func (m *Category) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []Category, total int64, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
