+++ models/dict_data.go
@@ -102,7 +102,7 @@ func (m *DictData) FindById(id int) (dictData DictData, err error) {
 	return
 }
 
-func (m *DictData) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []DictData, total int, err error) {
+func (m *DictData) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []DictData, total int64, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
