+++ models/dict_type.go
@@ -100,7 +100,7 @@ func (m *DictType) FindById(id int) (dictType DictType, err error) {
 	return
 }
 
-func (m *DictType) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []DictType, total int, err error) {
+func (m *DictType) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []DictType, total int64, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
