+++ models/post.go
@@ -101,7 +101,7 @@ func (m *Post) FindById(id int) (post Post, err error) {
 	return
 }
 
-func (m *Post) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []Post, total int, err error) {
+func (m *Post) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []Post, total int64, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
