+++ models/role.go
@@ -100,7 +100,7 @@ func (m *Role) FindById(id int) (role Role, err error) {
 	return
 }
 
-func (m *Role) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []Role, total int, err error) {
+func (m *Role) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []Role, total int64, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
